Move ReasonFilterOptions into a new folder
